고정변수생성 클래스 추가 ,테스트 수정, 리팩터링
