--- src/main/java/Domain/GameInformation.java
@@ -1,20 +1,0 @@
-package Domain;
-
-public class GameInformation {
-
-    private final PlayerNames playerNames;
-    private final PrizeNames prizeNames;
-
-    public GameInformation(PlayerNames playerNames, PrizeNames prizeNames) {
-        this.playerNames = playerNames;
-        this.prizeNames = prizeNames;
-    }
-
-    public PlayerNames getPlayerNames() {
-        return playerNames;
-    }
-
-    public PrizeNames getPrizeNames() {
-        return prizeNames;
-    }
-}
