--- src/main/java/LadderDomain/BooleanValueGenerator.java
@@ -1,5 +1,0 @@
-package LadderDomain;
-
-public interface BooleanValueGenerator {
-    boolean generate();
-}
