--- src/main/java/LadderDomain/ConnectionStatus.java
@@ -1,17 +1,0 @@
-package LadderDomain;
-
-public enum ConnectionStatus {
-    CONNECTED,
-    DISCONNECTED;
-
-    public boolean isConnected() {
-        return this == CONNECTED;
-    }
-
-    public static ConnectionStatus from(boolean connect) {
-        if (connect) {
-            return CONNECTED;
-        }
-        return DISCONNECTED;
-    }
-}
