--- src/main/java/LadderDomain/LadderConnectionGenerator.java
@@ -1,13 +1,0 @@
-package LadderDomain;
-
-import java.util.Random;
-
-public class LadderConnectionGenerator implements BooleanValueGenerator {
-
-    private final Random random = new Random();
-
-    @Override
-    public boolean generate() {
-        return random.nextBoolean();
-    }
-}
