+++ src/main/java/domain/PrizeNames.java
@@ -1,0 +1,24 @@
+package domain;
+
+import java.util.List;
+
+public class PrizeNames {
+
+    private final List<PrizeName> prizeNames;
+
+    public PrizeNames(List<PrizeName> prizeNames) {
+        this.prizeNames = List.copyOf(prizeNames);
+    }
+
+    public int size() {
+        return prizeNames.size();
+    }
+
+    public PrizeName get(int index) {
+        return prizeNames.get(index);
+    }
+
+    public List<PrizeName> getAll() {
+        return prizeNames;
+    }
+}
