+++ src/main/java/ladderdomain/BooleanValueGenerator.java
@@ -1,0 +1,6 @@
+package ladderdomain;
+
+@FunctionalInterface
+public interface BooleanValueGenerator {
+    boolean generate();
+}
