+++ src/main/java/ladderdomain/LadderConnectionGenerator.java
@@ -1,0 +1,13 @@
+package ladderdomain;
+
+import java.util.Random;
+
+public class LadderConnectionGenerator implements BooleanValueGenerator {
+
+    private final Random random = new Random();
+
+    @Override
+    public boolean generate() {
+        return random.nextBoolean();
+    }
+}
